+++ src/vm/runners/builtin_runner/mod.rs
@@ -79,6 +79,8 @@ impl BuiltinRunner {
         }
     }
 
+    // Important note: the second returned value corresponds to the builtin's stop_ptr, which must be updated after calling this method
+    // It is not updated inside this method due to mutability problems
     pub fn final_stack(
         &self,
         vm: &VirtualMachine,
@@ -188,7 +190,7 @@ impl BuiltinRunner {
         Ok((0..segment_size).map(|i| (base, i).into()).collect())
     }
 
-    pub fn get_memory_segment_addresses(&self) -> (&'static str, (isize, Option<usize>)) {
+    pub fn get_memory_segment_addresses(&self) -> (isize, Option<usize>) {
         match self {
             BuiltinRunner::Bitwise(ref bitwise) => bitwise.get_memory_segment_addresses(),
             BuiltinRunner::EcOp(ref ec) => ec.get_memory_segment_addresses(),
@@ -977,31 +979,19 @@ mod tests {
     fn get_memory_segment_addresses_test() {
         let bitwise_builtin: BuiltinRunner =
             BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default(), true).into();
-        assert_eq!(
-            bitwise_builtin.get_memory_segment_addresses(),
-            ("bitwise", (0, None)),
-        );
+        assert_eq!(bitwise_builtin.get_memory_segment_addresses(), (0, None),);
         let ec_op_builtin: BuiltinRunner =
             EcOpBuiltinRunner::new(&EcOpInstanceDef::default(), true).into();
-        assert_eq!(
-            ec_op_builtin.get_memory_segment_addresses(),
-            ("ec_op", (0, None)),
-        );
+        assert_eq!(ec_op_builtin.get_memory_segment_addresses(), (0, None),);
         let hash_builtin: BuiltinRunner = HashBuiltinRunner::new(8, true).into();
-        assert_eq!(
-            hash_builtin.get_memory_segment_addresses(),
-            ("pedersen", (0, None)),
-        );
+        assert_eq!(hash_builtin.get_memory_segment_addresses(), (0, None),);
         let output_builtin: BuiltinRunner = OutputBuiltinRunner::new(true).into();
-        assert_eq!(
-            output_builtin.get_memory_segment_addresses(),
-            ("output", (0, None)),
-        );
+        assert_eq!(output_builtin.get_memory_segment_addresses(), (0, None),);
         let range_check_builtin: BuiltinRunner =
             BuiltinRunner::RangeCheck(RangeCheckBuiltinRunner::new(8, 8, true));
         assert_eq!(
             range_check_builtin.get_memory_segment_addresses(),
-            ("range_check", (0, None)),
+            (0, None),
         );
     }
 
@@ -1020,7 +1010,7 @@ mod tests {
             true,
         ));
         let vm = vm!();
-        // Unsed builtin shouldnt fail security checks
+        // Unused builtin shouldn't fail security checks
         assert_eq!(builtin.run_security_checks(&vm), Ok(()),);
     }
 
@@ -1488,7 +1478,7 @@ mod tests {
 
         for mut br in builtins {
             br.set_stop_ptr(ptr);
-            let (_, (_, stop_ptr)) = br.get_memory_segment_addresses();
+            let (_, stop_ptr) = br.get_memory_segment_addresses();
             assert_eq!(stop_ptr, Some(ptr));
         }
     }